requirements.txt
pip==21.2.4
python-dotenv==0.19.0
setuptools==49.2.1
slack-bolt==1.8.1
slack-sdk==3.10.1
